Mysql: use Exec_Master_Log_Pos
When we restore this snapshot, we tell mysql "This is the position we
were at" and that corresponds with "Exec_Master_Log_Pos" (the pos of
the last executed SQL command). "Read_Master_Log_Pos" is the position of
the last entry written to the relay log which we dont' care about since
we're not saving the relay log and not setting it's position on restore.